Tests with browserstack with different browsers
Related to: AAC-889, AAC-893
