+++ Resources/Private/Scss/_textmedia.scss
@@ -11,7 +11,7 @@ p {
 
 figure {
     margin: 0;
-    display: flex;
+    display: block;
     margin-top: $magic-l/3;
     flex-grow: 0;
     flex-wrap: wrap;
@@ -22,7 +22,7 @@ figure {
         font-weight: normal;
         font-size: $small-font-size-l;
         line-height: $small-line-height-l;
-        display: table-caption;
+        display: inline-block;
         caption-side: bottom;
         justify-content: flex-end;
         flex-direction: column;
