+++ Resources/Private/Scss/extensions/_pazpar2.scss
@@ -160,12 +160,15 @@
 
         ul {
             display: flex;
+            flex-direction: column;
             padding: 0;
 
             li input {
                 border: 0 none;
                 background: transparent;
                 cursor: pointer;
+                display: flex;
+                padding: 0;
 
                 input,
                 a {
