+++ Tests/output/.directory
@@ -1,0 +1,4 @@
+[Dolphin]
+PreviewsShown=true
+Timestamp=2018,2,27,12,1,4
+Version=3
