+++ codecept.json-browserstack
@@ -1,0 +1,30 @@
+{
+    "output": "test/output",
+    "helpers": {
+        "WebDriverIO": {
+            "smartWait": 10000,
+            "restart": false,
+            "url": "http://localhost:8001",
+            "browser": "firefox",
+            "user": "sibyllengle1",
+            "key": "szHA1pyYpQZx8VN34JKt",
+            "desiredCapabilities": {
+                "browserstack.local": "true",
+                "browserstack.debug": "true"
+            }
+        },
+        "GeneratedHelper": {
+            "require": "./test/PositionHelper.js"
+        }
+    },
+    "include": {
+        "I": "./test/libs/steps_file.js"
+    },
+    "mocha": {},
+    "bootstrap": false,
+    "teardown": null,
+    "hooks": [],
+    "tests": "test/*.test.js",
+    "timeout": 10000,
+    "name": "zivver-webapp"
+}
