--- test/steps_file.js
@@ -1,12 +1,0 @@
-
-'use strict';
-// in this file you can append custom step methods to 'I' object
-
-module.exports = function() {
-  return actor({
-
-    // Define custom steps here, use 'this' to access default methods of I.
-    // It is recommended to place a general 'login' function here.
-
-  });
-};
